Clean up and get distributed sort to compile

+++ src/enclave/Enclave/Project.cpp
@@ -4,8 +4,8 @@
 #include "common.h"
 
 void project(uint8_t *project_list, size_t project_list_length,
-             uint8_t *input_rows, uint32_t input_rows_length,
-             uint8_t **output_rows, uint32_t *output_rows_length) {
+             uint8_t *input_rows, size_t input_rows_length,
+             uint8_t **output_rows, size_t *output_rows_length) {
   flatbuffers::Verifier v(project_list, project_list_length);
   check(v.VerifyBuffer<tuix::ProjectExpr>(nullptr),
         "Corrupt ProjectExpr %p of length %d\n", project_list, project_list_length);
